update migrations and transaction

+++ src/conf/migrations/1659422720397_add-column-createdAt-and-updatedAt-into-products-table.js
@@ -1,7 +1,7 @@
 module.exports = {
   "up": `ALTER TABLE products
-          ADD createdAt INT NOT NULL AFTER categoryId,
-          ADD updatedAt INT NOT NULL AFTER createdAt`,
+          ADD createdAt BIGINT(32) NOT NULL AFTER categoryId,
+          ADD updatedAt BIGINT(32) NOT NULL AFTER createdAt`,
   "down": `ALTER TABLE products
             DROP createdAt,
             DROP updatedAt`
